requirements.txt
aiosmtplib==3.0.2
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
bcrypt==4.2.0
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.7
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.1
fastapi-mail==1.5.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2==2.9.10
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyotp==2.9.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-jose==3.5.0
qrcode==8.2
redis==6.4.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
